+++ apps/playground/src/app/components/forms/toggle-switch/toggle-switch.component.ts
@@ -1,0 +1,7 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'app-toggle-switch',
+  templateUrl: './toggle-switch.component.html',
+})
+export class ToggleSwitchComponent {}