+++ libs/components/indicators/src/lib/modules/key-info/key-info.module.ts
@@ -1,5 +1,6 @@
 import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
+import { SkyTrimModule } from '@skyux/core';
 
 import { SkyKeyInfoLabelComponent } from './key-info-label.component';
 import { SkyKeyInfoValueComponent } from './key-info-value.component';
@@ -11,7 +12,7 @@ import { SkyKeyInfoComponent } from './key-info.component';
     SkyKeyInfoLabelComponent,
     SkyKeyInfoValueComponent,
   ],
-  imports: [CommonModule],
+  imports: [CommonModule, SkyTrimModule],
   exports: [
     SkyKeyInfoComponent,
     SkyKeyInfoLabelComponent,