+++ libs/sdk/e2e-schematics/src/generators/pr-comment/schema.d.ts
@@ -1,0 +1,11 @@
+export interface PrCommentGeneratorSchema {
+  pr: string;
+  storybooks: string;
+  apps?: string;
+}
+
+export interface PrCommentGeneratorOptions {
+  pr: string;
+  storybooks: string[];
+  apps: string[];
+}