+++ libs/sdk/e2e-schematics/src/generators/story/schema.d.ts
@@ -1,0 +1,7 @@
+export interface ComponentGeneratorSchema {
+  name: string;
+  module?: string;
+  project?: string;
+  cypressProject?: string;
+  generateCypressSpecs: boolean;
+}