+++ libs/sdk/e2e-schematics/src/utils/utils.ts
@@ -1,0 +1,12 @@
+import { decamelize } from '@nrwl/workspace/src/utils/strings';
+
+export function dirname(path: string): string {
+  const end = path.lastIndexOf('/');
+  return path.substring(0, end);
+}
+
+export function capitalizeWords(string: string) {
+  return decamelize(string)
+    .replace(/[-_]/g, ' ')
+    .replace(/\b[a-z]/g, (char) => char.toUpperCase());
+}